+++ src/main/java/com/example/food/model/Wallet.java
@@ -1,12 +1,13 @@
 package com.example.food.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jdk.jfr.Timestamp;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -20,7 +21,7 @@ public class Wallet {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long walletId;
-    private Integer walletBalance;
+    private BigDecimal walletBalance;
     @CreationTimestamp
     private Date createdAt;
     @UpdateTimestamp
